Refactor integration to its own directory
